requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=23.5
python-decouple>=3.8